WIP refactoring to match 4.2 structure
