+++ .gitignore
@@ -5,4 +5,4 @@ tmp/
 Gemfile.lock
 /gemfiles/*.gemfile.lock
 /travis/*.lock
-/test/database_local.yml
+/test/database.yml
