+++ activerecord-postgis-adapter.gemspec
@@ -20,5 +20,4 @@ require './lib/active_record/connection_adapters/postgis/version.rb'
   spec.add_dependency 'rgeo-activerecord', '~> 1.0'
 
   spec.add_development_dependency 'rake', '~> 10.2'
-  spec.add_development_dependency 'mocha', '~> 1.0'
 end
