--- lib/active_record/connection_adapters/postgis/setup.rb
@@ -1,11 +1,0 @@
-module ActiveRecord  # :nodoc:
-  module ConnectionAdapters  # :nodoc:
-    module PostGIS  # :nodoc:
-
-      def self.initial_setup
-        ::ActiveRecord::SchemaDumper.ignore_tables |= %w[geometry_columns spatial_ref_sys layer topology]
-      end
-
-    end
-  end
-end
