+++ test/database.example.yml
@@ -1,0 +1,13 @@
+default_env:
+  adapter: postgis
+  host: 127.0.0.1
+  database: postgis_adapter_test
+  username: postgres
+  schema_search_path: public
+
+default_env2:
+  adapter: postgis
+  host: 127.0.0.1
+  database: postgis_adapter_test
+  username: postgres
+  schema_search_path: public
