--- test/database.yml
@@ -1,7 +1,0 @@
-adapter: postgis
-host: 127.0.0.1
-database: postgis_adapter_test
-username: abdelkader
-password: q
-setup: default
-schema_search_path: public
